moved main components to a subfolder
