+++ src/components/main/Main/index.js
@@ -2,13 +2,13 @@ import React from "react";
 import { makeStyles, Grid } from "@material-ui/core";
 import { Route, Switch } from "react-router";
 import Logout from "components/auth/Logout";
-import Profile from "components/Profile";
+import Profile from "components/main/Profile";
 import Admin from "components/admin/Home";
 import NewCampaign from "components/intel/NewCampaign";
 import Intel from "components/intel/Home";
 import NewPin from "components/intel/NewPin";
 import PinHistory from "components/intel/PinHistory";
-import Home from "components/Home";
+import Home from "components/main/Home";
 import Unconfirmed from "components/auth/Unconfirmed";
 import Login from "components/auth/Login";
 import Register from "components/auth/Register";
